Time: 2 ms (86.21%), Space: 40.3 MB (14.98%) - LeetHub

+++ 0441-arranging-coins/0441-arranging-coins.java
@@ -2,13 +2,13 @@ class Solution {
     public int arrangeCoins(int n) {
         int left = 1, right = n, max = 0;
         while (left <= right) {
-            int mid = left + (right - left) / 2;
-            long cnt = (long)((1 + mid) * (mid / 2.0));
-            if (cnt == n) return mid;
+            long mid = left + (right - left) / 2;
+            long cnt = ((1 + mid) * mid)/ 2;
+            if (cnt == n) return (int)mid;
             else if (cnt < n) {
-                left = mid + 1;
-                max = Math.max(max, mid);
-            }else right = mid - 1;
+                left = (int)mid + 1;
+                max = Math.max(max, (int)mid);
+            }else right = (int)mid - 1;
         }
         return max;
     }
